Move collection related view functions to their own file.
